Delta versioon 3.3.2.13
Selle versiooniga tulevad kaasa järgmised täiendused ja parandused:
* PPA102: TÜ täitja mitmekordselt sisestamise kontroll
* PPA098: Allkirjastamiseks käivitatud tööülesanded ei jõua täitjani
* T024-98 Ei saa automaatselt dokumendi sees allkirjastajaks määratud isikut allkirjastamise töövoos muuta
* T024-99 Ei saa mittekohustuslikult dokumenti sisestatud allkirjastajat töövoos allkirjastamiseks muuta
* T024-63 Süsteemi viga kasutajate halduses nime valikul
* T024-94 Malli ei ole võimalik üles laadida
* T024-100 Teostab toimingu, kuid kuvab veateate "Süsteemi viga"
* DLSeadist: Andmevälja detailvaates tüübiga Täisarv ei tagastata valideerimisel loetavat veateadet
* DLSeadist: Dokumendi liigi detailvaades, metaandmete plokkis kuvatav täiendav info ei ole korrektne
* DLSeadist: Dokumendi liikide import-eksport
* DLSeadist: Dokumendi liigi detailvaade - Lubatud seoste plokid
* DLSeadist: Andmevälja originalFieldId väärtusega arvestamine
* DLSeadist: Süsteemne dokumendi liik "Koolitustaotlus"
* DLSeadist: Süsteemne dokumendi liik "Leping"
* DLSeadist: Süsteemne dokumendi liik "Välislähetuse korraldus"
* DLSeadist: Süsteemne dokumendi liik "Aruanne"
* DLSeadist: Dokumendi liigi detailvaate salvestamine peab kasutaja jätma samasse vaatesse
* DLSeadist: Tühja andmeväljade grupi kinnitamisel kuvatakse vale veateade
* DLSeadist: Vigane käitumine, kui dokumendi liigi/andmevälja identifikaator ei valideeru
* DLSeadist: süsteemsed andmeväljade grupid Adressaat ja Lisaadressaat - kontaktide valimine vormile ei tööta korrektselt

Teadaolevad vead:
* Dokumendi liikide importimine ei toimi vastavalt vajadusele – rakendus jooksis kokku.
* Kasutajate, kontaktide otsingus on filtreerimise funktsionaalsus välja võetud – vajab rohkem uurimist.
* Dokumendi liigi seadistamisel "Salvesta" nuppu järel kuvatav teade ei vasta spetsifikatsioonile.
* Suletud sarjas kuvatakse toimiku detailvaates "Taasava" nuppu ja vajutamise järel kuvatatakse teade, mille sisu ei ole õige.


git-svn-id: https://svn.smit/tarkvara/edhs/trunk@109318 5fddf92a-9921-11dd-8d32-b999f448c456

+++ delta/common/source/java/ee/webmedia/alfresco/base/BaseServiceImpl.java
@@ -24,6 +24,7 @@ import org.alfresco.service.namespace.QNamePattern;
 import org.springframework.util.Assert;
 
 import ee.webmedia.alfresco.common.service.GeneralService;
+import ee.webmedia.alfresco.common.web.BeanHelper;
 import ee.webmedia.alfresco.common.web.WmNode;
 import ee.webmedia.alfresco.utils.MessageDataImpl;
 import ee.webmedia.alfresco.utils.Predicate;
@@ -35,7 +36,7 @@ import ee.webmedia.alfresco.utils.UnableToPerformException.MessageSeverity;
  * @author Alar Kvell
  */
 public class BaseServiceImpl implements BaseService {
-    private static final org.apache.commons.logging.Log log = org.apache.commons.logging.LogFactory.getLog(BaseServiceImpl.class);
+    private static final org.apache.commons.logging.Log LOG = org.apache.commons.logging.LogFactory.getLog(BaseServiceImpl.class);
 
     private DictionaryService dictionaryService;
     private NamespaceService namespaceService;
@@ -48,7 +49,10 @@ public class BaseServiceImpl implements BaseService {
     public void addTypeMapping(QName type, Class<? extends BaseObject> clazz) {
         Assert.notNull(type, "type");
         Assert.notNull(clazz, "class");
-        Assert.isTrue(!typeMappings.containsKey(type), "type is already mapped");
+        if (!BeanHelper.getApplicationService().isTest()) {
+            // this check is disabled in development to allow JRebel do it's magic when reloading spring context
+            Assert.isTrue(!typeMappings.containsKey(type), "type is already mapped");
+        }
         typeMappings.put(type, clazz);
     }
 
@@ -142,8 +146,8 @@ public class BaseServiceImpl implements BaseService {
 
             boolean wasSaved = object.isSaved();
             if (!wasSaved) {
-                if (log.isTraceEnabled()) {
-                    log.trace("Creating node (type '" + node.getType().toPrefixString(namespaceService) + "') with properties " + WmNode.toString(props, namespaceService));
+                if (LOG.isTraceEnabled()) {
+                    LOG.trace("Creating node (type '" + node.getType().toPrefixString(namespaceService) + "') with properties " + WmNode.toString(props, namespaceService));
                 }
                 if (!isDuplicateChildNamesAllowed(object, parent)) {
                     // make sure that when model forbids duplicates, then cm:name property is set that is used to check for duplicates
@@ -157,8 +161,8 @@ public class BaseServiceImpl implements BaseService {
                 // removing aspects is not implemented - not needed
             } else {
                 if (!props.isEmpty()) {
-                    if (log.isTraceEnabled()) {
-                        log.trace("Updating node (type '" + node.getType().toPrefixString(namespaceService) + "') with properties " + WmNode.toString(props, namespaceService));
+                    if (LOG.isTraceEnabled()) {
+                        LOG.trace("Updating node (type '" + node.getType().toPrefixString(namespaceService) + "') with properties " + WmNode.toString(props, namespaceService));
                     }
 
                     nodeService.addProperties(node.getNodeRef(), props); // do not replace non-changed properties
@@ -177,8 +181,13 @@ public class BaseServiceImpl implements BaseService {
                 for (BaseObject removedChild : (List<? extends BaseObject>) removedChildListEntry.getValue()) {
                     WmNode removedNode = removedChild.getNode();
                     if (removedNode.isSaved()) {
-                        Assert.isTrue(wasSaved, "Node that was just created cannot contain removed child that was previously saved");
-                        nodeService.deleteNode(removedNode.getNodeRef());
+                        if (wasSaved) {
+                            nodeService.deleteNode(removedNode.getNodeRef());
+                        } else {
+                            if (LOG.isTraceEnabled()) {
+                                LOG.debug("not deleting node, that was previously saved under some other parent node: " + removedNode.getNodeRef());
+                            }
+                        }
                         changed = true;
                     }
                 }
@@ -207,7 +216,7 @@ public class BaseServiceImpl implements BaseService {
                 Integer assocIndex = entry.getKey();
                 for (BaseObject child : entry.getValue()) {
                     if (!wasSaved) {
-                        Assert.isTrue(!child.isSaved(), "Node that was just created cannot contain child that was previously saved");
+                        Assert.isTrue(child.isUnsaved(), "Node that was just created cannot contain child that was previously saved");
                         child.updateParentNodeRef();
                     }
                     boolean childChanged = saveObject(child);
